Move DropDownHolder into plugins folder to avoid require circles
